Remove more warnings (i386 build)

+++ libnux/framebuffer.c
@@ -70,9 +70,9 @@ framebuffer_blt (unsigned x, unsigned y, uint32_t color,
 	  while (bits)
 	    {
 	      if (byte & 0x80)
-		*(volatile uint32_t *) (fbdesc->addr + off) = color;
+		*(volatile uint32_t *) (uintptr_t) (fbdesc->addr + off) = color;
 	      else
-		*(volatile uint32_t *) (fbdesc->addr + off) = 0;
+		*(volatile uint32_t *) (uintptr_t) (fbdesc->addr + off) = 0;
 
 	      off += bypp;
 	      byte <<= 1;
@@ -132,7 +132,7 @@ framebuffer_putc (int ch, uint32_t color)
   if (!init)
     {
       spinlock_init (&fblock);
-      memset ((void *) fbdesc->addr, 0, fbdesc->size);
+      memset ((void *) (uintptr_t) fbdesc->addr, 0, fbdesc->size);
       __scrcol = (fbdesc->width / 8) / (__cols + 1);
       __scrcol = __scrcol == 0 ? 1 : __scrcol;
       __lines = fbdesc->height / 16;
